backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
pydantic-settings==2.8.0
email-validator==2.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
httpx==0.28.1
python-multipart==0.0.20
